requirements.txt
nltk==3.4.5
numpy==1.15.4
scipy==1.2.3
PyYAML==5.1
torch==0.4.1.post2
networkx==2.1
tensorflow==1.14.0
scikit-learn==0.20.0
argparse==1.4.0
